refactor(assets): extract asset-reader and asset-listing use-cases from the tool layer
The two asset handlers were the only handlers in the codebase
orchestrating multi-step pipelines inline (importing a parser into
mcp-core to do it). The dispatch and listing compositions move to
vault-operations use-case modules — the note-mover/vault-patcher
pattern — leaving the handlers the sibling silhouette: schema, wire
mapping, one data-layer call, formatting.

Along the way, per review: extension filter is a Set (O(1) membership),
extensionCounts is an honest immutable reduce (bounded accumulator width
— distinct extensions — so the O(n·k) concern that killed the file-list
reduce doesn't apply), and the misleading pagination vocabulary is gone
(pagePaths/stattedPage → returnedPaths/stattedAssets; there is no
pagination — limit caps the slice and counts/total tell the caller to
narrow). AGENTS.md codifies the thin-handler rule with these modules as
worked examples; structure tree, wiki.json, and ARCHITECTURE updated.

Behavior and wire contract unchanged — the tool-layer suite covers every
arm through the new modules.
